pyproject.toml
[project]
name = "github"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.99.9",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
